handle lack of mysql support correctly in backup, restore and rollback
logic
    backup
        if it can't open mysql.mysqldump()
            don't create myfs in extras

    restore
        if myfs doesn't exist or mysql.mysql() doesn't open
            doesn't restore mysql

        the same way it wouldn't restore pgsql if that didn't work

        why we shouldn't try to fix things
            an infinite number of things could go wrong
            let the user fix it
                e.g., install mysql

    rollback
        if rollback/myfs exists and we can't rollback mysql raise an Exception
            because we're not rolling back - that's an error we shouldn't ignore

+++ restore.py
@@ -75,10 +75,14 @@ class Restore:
         if self.rollback:
             self.rollback.save_database()
 
-        mysql.fs2mysql(mysql.mysql(), self.extras.myfs, self.limits.db, mysql.cb_print(self.log))
+        if exists(self.extras.myfs):
+            try:
+                mysql.fs2mysql(mysql.mysql(), self.extras.myfs, self.limits.db, mysql.cb_print(self.log))
+                shutil.copy(join(self.extras.etc.mysql, "debian.cnf"), "/etc/mysql/debian.cnf")
+                os.system("killall -HUP mysqld > /dev/null 2>&1")
 
-        shutil.copy(join(self.extras.etc.mysql, "debian.cnf"), "/etc/mysql/debian.cnf")
-        os.system("killall -HUP mysqld > /dev/null 2>&1")
+            except mysql.Error:
+                pass
         
     def packages(self):
         newpkgs_file = self.extras.newpkgs
